requirements.txt
argparse>=1.2
Flask
Flask-Login
Flask-Script
Flask-SQLAlchemy
odict>=1.5
prettytable>=0.7.2
requests>=2.4
requests-futures
six>=1.7


